refactor(TermsConditionsPage): Merge with main and fix to work
Merge with main and a few code refactor.

folders and files padronization
