+++ src/pages/termsConditions/components/form.jsx
