+++ src/pages/termsConditions/index.jsx
